+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 History
 -------
 
+ - Add parameter separator to function range_conv
+
 **2021-05-27 (0.2.0)**
  - Add overview table to API documentation
  - All parameters but 'string' are now keyword-only